requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.7
kml2geojson==4.0.2
Pillow==8.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
simplekml==1.3.5
six==1.15.0
sqlparse==0.4.1
gdal>=2.0.0
